+++ src/interfaces/command.ts
@@ -2,7 +2,6 @@ import { Message } from 'discord.js';
 
 export default interface Command {
     run: (message: Message, args: string[], messages: any) => any;
-
     info: {
         names: string[],
         description: string,